feat: implement file update functionality, refactor API response models, and standardize database ID types to Integer

+++ src/main/java/com/furniro/UploadService/database/repository/FileUploadRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FileUploadRepository extends JpaRepository<FileUpload, Long> {
+public interface FileUploadRepository extends JpaRepository<FileUpload, Integer> {
 }
